Temporarily use popup instead of redirect auth
Redirect was only sending the user:email scope to GitHub even when requesting repo. As per https://github.com/angular/angularfire2/issues/798 this does not happen when using popup. However, the accessToken isn't visible in the observable of auth states, but is in the promise returned from login.
Given the impending RC of AngularFire2 will require reworking of the auth code (see https://github.com/angular/angularfire2/issues/854), this will do for now.